Add type=module
Attempt to create modules
Known CORS errors
