+++ src/index.html
@@ -110,7 +110,7 @@
     <!-- SmtpJS -->
     <!-- <script src="https://smtpjs.com/v3/smtp.js"></script> -->
     <!-- XML -->
-    <script src="assets/js/app.js"></script>
+    <script type="module" src="assets/js/app.js"></script>
 </body>
 
 </html>
